Port sample-on-last-keyframe test to WPT
The non-WPT version of this test started flaking. The test used a
legacy API, so rather than fixing in place, it seemed like a good
opportunity to migrate the test to WPT and use a modern API.

The opacity of the end-keyframe was changed from 1 to differentiate
the animation being in effect from being idle.

Bug: 1348835
Change-Id: Iddd45e8fed793692a1121ae218fa470d0aa6bb98