[balance-text] Implement balancing algorithm
This patch implements `NGParagraphLineBreaker` that implements
the headline balancing algorithm. It computes the minimum
available width that produces the same number of lines as the
normal line breaker by bisecting.

The algorithm is O(log n). To mitigate the performance penalty,
it skips if the block has more than 4 lines, or it is
estimated to have more than 8 lines.

See the document[1] for more details.

[1] https://docs.google.com/document/d/16-T9gqCagJxcST6hcnneSb7qGunxXa37_UHYqMqhPL0/edit?usp=sharing

Bug: 1251079
Change-Id: I31a5859293622c14dda05e785c4495cd82db6f14