[Private Network Access] Test redirect from same origin to cross origin
At [1], a concern was raised about the implementation that allows secure
same-origin private network requests without preflights. There was a
possibility that redirects were not correctly handled.

The test cases added in this CL test redirects from same-origin to
cross-origin or vice versa.

Some failed tests in virtual/pna-allowed-secure-same-origin are
expected, but some also indicate bugs, which we'll try to fix in a later
CL.

This CL alsoremoves the UUID from the response body because it prevents
UUID from going into the expected output.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/4221658/comment/ca7f33a7_1599ae6f/

Bug: 1382068
Change-Id: I96352dcba7c7c6517aac0108d24e1fa0491aacbe